requirements-vercel.txt
Flask==2.3.3
opencv-python-headless==4.9.0.80
mediapipe==0.10.21
numpy==1.26.4
Pillow==10.2.0
Werkzeug==2.3.7 